big-data (Python and data analysis)/Dashboard/requirements.txt
flask==2.0.1
flask-sqlalchemy
pandas
numpy
mysqlclient